requirements.txt
pyobjc_core>=7.0.1
PyQt6>=6.2.0
